LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/intrinsic-size-fallback-video-expected.txt


PASS .wrapper 1
FAIL .wrapper 2 assert_equals:
<div class="wrapper" data-expected-client-width="300">
  <video style="height: 100px"></video>
</div>
clientWidth expected 300 but got 200
FAIL .wrapper 3 assert_equals:
<div class="wrapper" data-expected-client-width="300">
  <video style="max-height: 50px"></video>
</div>
clientWidth expected 300 but got 100
FAIL .wrapper 4 assert_equals:
<div class="wrapper" data-expected-client-width="300">
  <video style="min-height: 200px"></video>
</div>
clientWidth expected 300 but got 400
